move app related funcs to a separate appkit inside testkit
